backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
google-generativeai
python-dotenv

